STASHDEV-7788 Moved `executingCalls` and all logic implementing the `isOperationExecuting` method from `BasicOperationService` to `BasicOperationScheduler`.  This allows the "is-executing" remote operation to see invocations that are stuck in a work queue.
